requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
google-genai>=1.0.0
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow>=10.2.0
duckduckgo-search>=6.1.0
python-docx==1.1.0
pandas>=2.0.0
pydantic>=2.6.0
python-dotenv==1.0.0
aiofiles>=23.2.1
